+++ changelog.md
@@ -8,6 +8,7 @@
   absolute.
 - Add `magicalBraces` option to treat brace expansion as "magic"
   in the `hasMagic` function.
+- Add `dotRelative` option
 
 ## 9.0